+++ R/utils.R
@@ -307,3 +307,15 @@ infer_shapes = function(shapes_in, param_vals, output_names, fn, rowwise, id) {
 
   set_names(list(sout), output_names)
 }
+
+get_or_check_dataset_shapes = function(dataset, dataset_shapes) {
+  if (is.null(dataset_shapes)) {
+    if (is.null(dataset$.getbatch)) {
+      stopf("dataset_shapes must be provided if dataset does not have a `.getbatch` method.")
+    }
+    dataset_shapes = infer_shapes_from_getbatch(dataset)
+  } else {
+    assert_compatible_shapes(dataset_shapes, dataset)
+  }
+  dataset_shapes
+}